Debugging
Handled some bugs:
- Locate point was not finding points in odd-numbered faces, as
triangles were not being given the other triangle in the same square as
a neighbour and so these were not being explored.
- Barycentric coordinates were not being properly divided by the total
triangle area during computation, and some of the heights on the
vectors were wrong, causing false results.
- The contains method had not been completed yet, and was instead
hard-coded to return true (doh!)
And various other one-liners.